RPM: createrepo editor and project creation
Changed project creation to create and store the repo file if initial
creation, or find the repo file if existing project.

Added a listener to listen on when a createrepo project has been
deleted/closed to also close the relating file if open in an editor.

Change-Id: I175bc8831a2212cefc2e6f9ae01797a935323344
Signed-off-by: Neil Guzman <nguzman@redhat.com>
Reviewed-on: https://git.eclipse.org/r/18849
Tested-by: Hudson CI
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
IP-Clean: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>