Systemtap: Pre-run autosave for projectless scripts.
When a Systemtap script that lacks a parent project is run,
have that script be immediately saved before the run begins.

This adds to the functionality of fa6613e, which only saves
scripts that belong to a project. This improvement also fixes a
bug introduced in that patch which caused a NullPointerException
when running a projectless script.

This patch also fixes a bug in fa6613e that allowed non-*.stp
files to be run as Systemtap scripts.

Change-Id: Ib0cc475e882ad9775fdaca44b1119089ca65638c
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/19084
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
IP-Clean: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>
Reviewed-on: https://git.eclipse.org/r/19159
Tested-by: Hudson CI