ctf: Fix decoding of an invalid packet id type.
When an event id is not present, it is assumed to be zero.
Otherwise, it must be a _valid_ integer.

The current implementation accepts a "struct" as an id and map to the event id zero.

This fix throw an exception on an invalid id.

Change-Id: Iff6f6bd0173048dd3fb9b00f564948d7ebd55f4f
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/19166
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI