ctf: Make CtfTmfEventFieldTest easier to maintain.
Before this patch, a test event was created by filling a buffer with 0x02
bytes. Then, 0x00 bytes were added at magic positions to terminate the
strings used in the tests. That made the code hard to read and maintain. In
particular, it was hard to add new values in the test event without breaking
the null-terminated strings. It was also hard to understand, for example, that
0x02020202 is the float 9.551467814359616E-38.

With this patch, values in the right format are added explicitly to the test
event. It's easier to understand the tests and to add values for new tests.

Change-Id: I5d482442a10f09627b3ed2b32a9a8cc68c0f5cdc
Signed-off-by: Francois Doray <fdoray.eclipse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19192
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
IP-Clean: Matthew Khouzam <matthew.khouzam@ericsson.com>
Tested-by: Matthew Khouzam <matthew.khouzam@ericsson.com>