tmf: Avoid HistoryTree to expose its internal HT_IO node.
The HT_IO node is an abstraction of the file accesses for
a tree node. It should not be exposed to others classes.
The history tree must act as a proxy for requests to HT_IO.
This will ease modifications and maintenance of HT_IO.

HT_IO node should not keep a back reference to the HistoryTree.
This is a bad design. HT_IO node needs to know about HTConfig
to perform its tasks.

The HT_IO class contains method to read from disk.
The concept of a LastBranch is related to the HistoryTree.

Thus, readFromMemory is the same than calling

  HistoryTree.readNode(...)

and readFromDisk is the same than calling

  HT_IO.readNode(...)

Each layer is reponsible of its own caching mechanism.

Change-Id: Ia835a03efa321cdefd170de002833e044d144ee0
Signed-off-by: Etienne Bergeron <etienne.bergeron@gmail.com>
Reviewed-on: https://git.eclipse.org/r/19557
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>