ctf: Generate synthetic trace in test traces folder instead of temp dir
I was having problems with the synthetic trace now that the Luna branch
is merged into master because the index files got generated into the
synthetic trace folder and the folder would not get deleted on exit
because of that. Then the next test run, the trace would not get
generated because the folder already exists. I think it makes sense to
move the trace to the test traces folder and only generate it once.

Change-Id: I03e815c50e22adf2d5355856b8031cc7a8b4c9b1
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/19680
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
IP-Clean: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Tested-by: Hudson CI