TMF: Fix a bug in UI when refreshing analysis elements under traces
As it was, each time we would get the available analysis of a trace, it was
constructing new objects and those objects were not the same as the trace
element's children.

Also, while implementing the XML analysis source, the analysis were not
refreshed when the trace didn't already have children.

This patch uses the refreshChildren method to refresh analysis and outputs and
makes new objects are not constructed every time.

Change-Id: Ia33f7e20fea918a76bb341b808d074d3b1e130f2
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/20004
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
IP-Clean: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Hudson CI