tmf: Work around SWTBot test hang in the presence of GTK3
With GTK3, the SWTBot test hangs as soon as a Display.syncExec
is called on the swtbot thread (non-UI) which seems like a bug
in SWT. For now, GTK3 can be disabled by setting the SWT_GTK3
environment variable to 0. Note that for Kepler GTK2 is still
used by default so this is not a problem.

Change-Id: I4a5026f616ab032d95e90e4ad80cea665d552d3c
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/20407
Tested-by: Hudson CI