Systemtap: Fix/improve data filter UI.
Make the following changes to DataGrid's filter UI:

-Add names to filters when they are shown in the "remove
filters" menu, and disable that menu when there are no
filters.

-Implement singly-bounded range filters.

-Prevent bugs that are caused by trying to get data from
an empty table/data set.

-In the Select Filter wizard, only create a filter if
"Finish" is clicked when a wizard page with a valid
filter is the current page.

-Always cause a full update to the DataGrid on each update
event, as filters may change the order of table entries.

Change-Id: If395acbeb5aa77e96b638e7c1ae227c8224589ce
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/21221
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>