tmf: support for renaming, overwriting or skipping during import
If a trace with a given name already exists in the target project
and the checkbox "Overwrite existing trace without warning" is not
checked, then a dialog box will be opened so that the user can
specify if he wants to rename, overwrite or skip the trace. The user
also has the option to apply this for all the imported traces, for
example "Rename all". For the renaming case the import operation will
automatically add a number as suffix to the trace name, for example
(2). This is similar to how it is done for drag and drop or
"Open Trace File/Directory" command in the RCP.

Change-Id: I316752fb7a76f6f448958b147e212af9a1ee2bcb
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/21832
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>