TMF: Added the possibility to define analysis outputs in extension point
A new element is added to the o.e.l.tmf.core.analysis extension point to be
able to associate an output to either an analysis (identified by ID) or an
analysis class. That will allow the analysis to be in the .core plugins and
views to be registered to them in the .ui plugins.

Change-Id: I7a3e542c250c35a95bc9c3712e14e731e30c5d86
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/21578
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>