TMF: Allow XY line chart classes to change the resolution of the viewer
While a resolution of 1 point per pixel shows a good level of precision for
charts showing absolute values at time t (like memory usage charts), it is not
the case when the data to show is a marginal difference with preceding point
(like CPU usage view). In those cases, a lower resolution makes better looking
graphs, so the viewer should be able to set the resolution.

Change-Id: I7cad9a85555f57fcb812dd774929925125b39ca0
Signed-off-by: Geneviève Bastien <gbastien+lttng@versatic.net>
Reviewed-on: https://git.eclipse.org/r/22102
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>