tmf: Replace "find" with "matches" for TextTrace.java
This means that if a pattern is not 100% correct, it will
fail to read an event. However, performance should increase.

If we have for example a pattern (\\d\\d)\s(\S+)
and "123 blabla" will work with find(), returning "23 blabla"
but will fail with matches().

Change-Id: I54e259f31b44fa2996316150665e5ca011576825
Signed-off-by: Matthew Khouzam <matthew.khouzam@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/22539
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
IP-Clean: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>