Systemtap: Fix "View Definition" (Function view).
Right-clicking a function in the Function view now correctly
displays the "View Definition" command in a context menu, and
the command's functionality has been restored (it now opens the
file where the chosen function is defined).

Note that the "View Definition" command has not yet been restored
for probes in the Probe Alias.

Change-Id: Ia431b56c87f2487f92bcc0b6b1437e316e643118
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/22699
Reviewed-by: Sami Wagiaalla <swagiaal@redhat.com>
IP-Clean: Sami Wagiaalla <swagiaal@redhat.com>
Tested-by: Sami Wagiaalla <swagiaal@redhat.com>