tmf: Fix NPE for persitent experiment index when using text traces
For text traces the end of a trace is indicated by a NULL_LOCATION.
The NULL_LOCATION has a null pointer for the location info object
(=TmfLongLocation). In an experiment some traces are finished being
read before other traces. When a checkpoint needs to be created
with a NULL_LOCATION for some of the traces an NullPointerExeption
is generated when serializing a checkpoint to the persistent index.

Change-Id: I16aa3afd1d7ab633b53916f7f3a01aefa9094560
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/22846
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
IP-Clean: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>