tmf: remove open trace directory and use open trace file instead
For directory traces like CTF the user has to select one of the
files of the directory trace. The open procedure will first verify
that the parent directory is a valid directory trace. If yes, then it
will be opened with the corresponding trace type. To cover the index
file of CTF which is in a subdirectory of the trace directory, also
the grandparent directory of the selected file will be checked for
a directory trace type. If neither parent or grandparent directory
is a valid directory the file will be imported as before.

Change-Id: Id6b4d8b177875b5c018cbbed82816f3ba52b067a
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/23386
Tested-by: Hudson CI
Reviewed-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Tested-by: Genevieve Bastien <gbastien+lttng@versatic.net>
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>