tmf: Refactor bookmarks file handling
The interfaces used to set or get the bookmarks file from an editor or
experiment are removed as they are not necessary. The bookmarks file is
the file in the FileEditorInput of an opened trace editor.

A convenience method is added in the trace manager to get the editor
file for an opened trace.

Change-Id: Ia4bd1999a4015b0251f867d46d0bd2d732a8c4bd
Signed-off-by: Patrick Tasse <patrick.tasse@gmail.com>
Reviewed-on: https://git.eclipse.org/r/24355
Tested-by: Hudson CI