tmf: Copy ResourceTreeAndListGroup from platform to keep compatibility
ResourceTreeAndListGroup was internal in Kepler and we referenced it.
It is now removed in Luna. To keep our builds compatible with Kepler,
we need to have our own version of this class. Once we stop supporting
Kepler, we can delete this class and use the public one from the platform.

Change-Id: I35185bbc9de551cacd6aaca3d12952f04c80d34e
Signed-off-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/26165
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>
Tested-by: Bernd Hufmann <bernd.hufmann@ericsson.com>