Gcov: Log test job failures.
Within a GCovTest, when a JobsRunning wait condition times out,
print information about the failure to standard output.
Also, increase the allowable timeout of the "dump CSV" job.

NOTE: the purpose of this patch is to aid in finding the cause
behind test failures, so the changes made to JobsRunning.java
should be reverted once the failures are fixed.

Change-Id: I4b4649003873ec19fe7ab46f63f2caeb1bf144f4
Signed-off-by: Andrew Ferrazzutti <aferrazz@redhat.com>
Reviewed-on: https://git.eclipse.org/r/26297
Reviewed-by: Alexander Kurtakov <akurtako@redhat.com>
Tested-by: Alexander Kurtakov <akurtako@redhat.com>