pcap: make PacketStream compute useful information
The PacketStream now computes Nb of packets/bytes, duration of
the conversation, Bytes per seconds, etc. Also, it now does not
store indexes to packet, since it makes the size of PacketStream
proportional to the size of the trace. This is unwanted since we
have to be able to open traces that are bigger than memory.

Change-Id: I8f6bde99ceabc20c4a66786a2dbd25d49251cc9e
Signed-off-by: Vincent Perot <vincent.perot@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/30998
Tested-by: Hudson CI
Reviewed-by: Alexandre Montplaisir <alexmonthy@voxpopuli.im>
Reviewed-by: Matthew Khouzam <matthew.khouzam@ericsson.com>