tmf: fix concurrency issue in analyses modules (bug 447434)
The access of member fTrace in class TmfAbstractAnalysesModul was
not thread-safe. The attribute was set to null when cancelling
the analyses (e.g. in the UI thread).

With this patch this attribute will be accessed (read and write)
using a synchronization lock and it will be checked for null
references before using it.

Additionally, this patch takes care of count down initialization
latches in case of cancellation.

Change-Id: If7a7c7685e4123747357b9fe8040a599ccc95c3a
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/34952
Reviewed-by: Patrick Tasse <patrick.tasse@gmail.com>
Tested-by: Hudson CI
Reviewed-by: Bernd Hufmann <bernd.hufmann@ericsson.com>