[lttng control] Use List<String> for control commands (Bug 455382)
By doing this, it's not necessary to surround commands with wildcards
and filters with quotes which made these commands fail using
org.eclipse.remote/Built-in SSH and org.eclipse.remote.Local.

For the PTP RSE adapter for org.eclipse.remote, the user has to enter
quotes otherwise the command will fail. The UI was changed to allow
quotes in the text boxes.

Change-Id: I2185c5b658f635ff491a6022fab206c208ca5d8f
Signed-off-by: Bernd Hufmann <Bernd.Hufmann@ericsson.com>
Reviewed-on: https://git.eclipse.org/r/38578
Tested-by: Hudson CI
Reviewed-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>
Tested-by: Marc-Andre Laperle <marc-andre.laperle@ericsson.com>