Bug 467885: Do not prepend another schema if one already exists.
When dealing with a unix/tcp socket, we would prepend the corresponding
schema if it was not found in the path. However in the case of tcp, the
schema is changed, causing us to prepend yet another schema when
reloading the connection.  Fix these kinds of issues by not prepending
any new schema if one is already present.

Change-Id: I705c6f6f5cc9592ecc17d991faf48acf3a0e31b6
Reviewed-on: https://git.eclipse.org/r/48384
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>