Bug 475881 - Pull image workflow improvement
Restored the PullImage wizard as the first step when
pulling an image.
The "Search..." buttons opens the SearchImage dialog with
search term prefilled from the current image name.
If no tag is specified, then 'latest' is assumed.
The validation only raises warnings to avoid preventing
the user to attempt to pull an image ( the Docker daemon
may eventually raise an error).

Added some JUnit tests to validate the image name patterns

Change-Id: I7b7fef9e8ed9bbbc7701880af44a4e9f412a1543
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/55494
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>
Tested-by: Roland Grunberg <rgrunber@redhat.com>