The VM creation wizard should support using box files.
Instead of requiring that the user 'add' the box through the Box View UI
before being able to use it, automatically do this if the user specifies
a valid path to a box in the 'Box Name' field.

The only assumption required is that a box file is a readable file path
and ends in '.box'.

Change-Id: I027f9363303af9eb5b703f6142c2a4915c9aba33
Reviewed-on: https://git.eclipse.org/r/60656
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>