Fix delayed rendering of VM state in the context menu.
Start using listChanged() from IVagrantVMListener to be informed of
changes to the set of VMs.

When fill() updates the context menu to reflect the current VM state,
the displayed context menu is from the previous state. Opening the
context menu a second time fixes this. To work around this issue, we
call an update() within listChanged() so that fill() is called twice.

Change-Id: I7859af3a5732307fef719e00202e26039e5e72b6
Reviewed-on: https://git.eclipse.org/r/61274
Tested-by: Hudson CI
Reviewed-by: Roland Grunberg <rgrunber@redhat.com>