Bug 494555: Quote command arguments when entrypoint is not defined.
By default the entrypoint should be '/bin/sh -c' so the arguments to
this should be quoted to work correctly. If this is done in the
setCommand method of the Image Run model, then we just need to ensure
that entrypoints are set before setCommand is called.

Change-Id: Iacec3a8557cec5dc1118244328ddac1003e6a801
Reviewed-on: https://git.eclipse.org/r/74098
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
(cherry picked from commit 3ae78b60e4481422318546a2bd5f832d86726c4b)
Reviewed-on: https://git.eclipse.org/r/74154