Bug 493752 - Docker Machine preferences page causes all other Docker preference pages to be invalid
Preference page is now valid if path to Docker Machine and
VM driver are empty. If the Docker Machine directory is valid
but does not contain the expected 'docker-machine' executable,
a warning is displayed.
If the directory paths are invalid, an error is displayed.

Change-Id: I489fe188540f47328e030e0b228e48ef938bc9f6
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/74121
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
(cherry picked from commit 0eb102607073159b42c36893b9df80266af59c5b)
Reviewed-on: https://git.eclipse.org/r/74136