Bug 496922 - Improvements on the 'push image' dialog
Refactored the ImagePushPage to use JFace databinding
Replaced button with link to open Registry account prefs
Added options to tag with 'force' flag and to keep the
tagged image upon completion.

Refactored ImagePushPage and ImagePullPage to use the same
base WizardPage to share the UI part that provides the
registry selection and addition.

Added SWTBot tests, refactored RegistryAccountManager by
introducing a RegistryAccountStorageManager that can be
replaced during tests.

Changed the validation in the RegistryDialog to allow
for null username/password since this should not be
a requirement for pulling images (eg: DockerHub or
other public registries)

Checks the name of the image before tagging.
Checks if the selected registry provides auth before pushing.
Display push progress handlers in the Progress view only.

Change-Id: I682b2cdace11c7124be691186ed31ea41741e2ba
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/76610
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>