Bug 499854 - Connection is not added(same parameters, different name)
Refactored the validators in the NewDockerConnectionPage
by adding  a search for connection with the same Unix/TCP settings when
the name is changed.
The page cannot complete if the user gives a new name for a connection with
known Unix socket path or TCP host.

Added SWTBot tests.

Change-Id: I99a4987c8c2c4c80aeffc58085ac1e7c94df50c9
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/79747
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
(cherry picked from commit f107e191a5cbe83dccc25172d74676b325bfbcc7)
Reviewed-on: https://git.eclipse.org/r/79872