proxy: fix in RuntimeProcessFactory#whichCommand method
After executing the which command in the machine, the method
was checking if the process error stream was not null and if
it was not null, it was supposing that the command was not
installed. This checking can be error phrone because you can
get any errors from the machine bash or things like that and
it will suppose that the command is not installed.

As the method already check the process input stream to get the
command path (or empty if not found), I am removing the error stream
checking.

Change-Id: If454b06596fc2915f034493440f18f5d8cbe8796
Signed-off-by: Roberto Oliveira <rdutra@linux.vnet.ibm.com>
Reviewed-on: https://git.eclipse.org/r/82657
Tested-by: Hudson CI
Reviewed-by: Wainer dos Santos Moschetta <wainersm@linux.vnet.ibm.com>