Bug 507143 - DockerException in ProcessUtils$FluentProcessBuilder.<init>
Instead of logging an error when the user settings are invalid, the tooling
displays an error message and does not attempt to search for Docker Machines.

Also, refactored the code to use the ProcessLauncher utility class instead of the
ProcessUtils class that was deprecated (ProcessLauncher is a follow up for
the ProcessUtils).

Change-Id: I68df8625a4b636b78b8b37fb5f1f001f772d81e4
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/84563
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>