Bug 507769 - NullPointerException when the list of images is not loaded in the Run launch config
Main problem was that the result of the call to `getImages(force)` was not assigned to
the local `latestImages` variable which was set to `null` (as `this.images` would be
`null` at this stage). Thus, even though the list of images was correctly loaded, it
would return `null` on the first call.

Also, removed the use of `containersLoaded` and `imagesLoaded` fields, which can be
replaced with tests on the `containers` and `images` lists instead.

Change-Id: I64584fbee2da1dfa8d6146c9ecfd89b3f3623e34
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/85314
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>