Bug 508282 - Unable to list image tags on DockerHub
Refactored the code in the AbstractRegistry class to move the logic of
retrieving the tags by type of registry in private methods, and
implemented
the workflow to retrieve a Bearer token if needed when querying on
DockerHub.

Also, providing the resulting list of tags in a sorted, reverse order.

Added a few tests to verify the parsing of the 'Www-Authenticate'
response header,
and verify that the list of tags for an official image (eg: "php") and a
regular image (eg: "jboss/wildfly") can be retrieved. These tests are
temporarily ignored because they fail on Hudson (can the salve machine
connect to Docker Hub ?)

Also: refactored some loops to use the Stream API.

Change-Id: I60d7d71188e2e6e3d19a32e6fb1e53cb11605bf8
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/85887
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
(cherry picked from commit 1f01289c7cbcf30ef593a0ad06448ed3c4e984ac)
Reviewed-on: https://git.eclipse.org/r/85986