Bug 508655 - DockerException in DockerConnection.open
Removed the call to `Activator.log(e)` when a problem occurs
while testing the connection settings in the wizard. Instead,
the exception's message is displayed in the error dialog, which
is more explicit to the user (no need to look for the reason
in the Error Log). Same pattern is applied on both
`NewDockerConnection` page and `EditDockerConnection` page.

Also, fixed a few generics warnings in the `NewDockerConnectionPage`
class.

Change-Id: Iac9d1df4c5f1f61268317ce3461b90284c1d1144
Signed-off-by: Xavier Coulon <xcoulon@redhat.com>
Reviewed-on: https://git.eclipse.org/r/86346
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
(cherry picked from commit 78320adc0ea5c3c0d90b49973cff80374a5a8f73)
Reviewed-on: https://git.eclipse.org/r/86420