Bug 508858: Pass a blocking InputStream to the TarArchiveInputStream.
The input stream returned from copyContainer may be passing back
available bytes which may not default to be 0.  In this case, the
tar input stream may be reading a header record and treat this
incomplete record as EOF.  Also fixed buffering to only write out
what was read (number of bytes).

Change-Id: Icbab8dc460eb613acf6e12cd2f5aba192c52e774
Reviewed-on: https://git.eclipse.org/r/86664
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>