Bug 517143: Prevent UI hangs due to reloadConnections in UI thread.
DockerConnectionManager.getInstance() will end up calling
reloadConnections() if the singleton has not yet been created. This can
happen when a user first opens the Docker Tooling perspective due to it
appearing in various createControl() methods. This code can take a few
seconds to execute so it should be done outside of the UI thread.

Change-Id: If62479fb88ab96f6d5d99bc151688b7c41308793
Reviewed-on: https://git.eclipse.org/r/97791
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>