Bug 530622: If isValid() is true, error messages should be cleared.
- When the Profiler tab is no longer active (clicked elsewhere) and then
becomes active once again we attempt to reload widget states from the
config. While this happens we are unable to make a determination for the
config's validity. Currently we return true for isValid() if not fully
initialized. However, the Profiler sub-tabs should have been loaded so
we can pass their isValid() methods the current config to make a
determination.

Change-Id: I025dbc44b1ed4cd8ee58d182728ae64d7f207740
Reviewed-on: https://git.eclipse.org/r/116553
Tested-by: Hudson CI
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>