Bug 566705 Docker: readContainerDirectory: Fix reading symlinks
"ls -L" reads info of linked file, not the link itself -> Links are
never marked as such -> Removed
The link target is now taken from ls and not normalized. The previous
solution also broke when links contained absolute paths. Normalizing
links also broke the PLA as, for example, extracting a Tar also does not
normalize symlinks.  File names and links with blanks in their names
are supported now by tokenizing properly.

ContainerFileProxy is updated to return the actual link via getLink()
and to return the actual target via getTarget()

CopyFromContainerCommandHandler is updated to create symbolic links
for relative links if copying a folder, otherwise, for individual files
and absolute links, the link is followed and the file/dir is copied.  Up
to 10 links are followed to avoid a possible circular link situation.

ContainerLauncher copy from container logic is updated similarly.

Fix API issue in DockerException caused by removal of getMessage().

Change-Id: Ie740e30ff1fbbf170e46d54eab134a82dc36c38b
Reviewed-on: https://git.eclipse.org/r/c/linuxtools/org.eclipse.linuxtools/+/173149
Tested-by: Linux Tools Bot <linuxtools-bot@eclipse.org>
Tested-by: Jeff Johnston <jjohnstn@redhat.com>
Reviewed-by: Jeff Johnston <jjohnstn@redhat.com>
Reviewed-on: https://git.eclipse.org/r/c/linuxtools/org.eclipse.linuxtools/+/168426